update models and migrations

+++ database/db-helper.js
@@ -1,5 +1,13 @@
 const { Player, CasinoGame, CasinoGameLog } = require('../database/models/index');
+const { v4: uuidv4 } = require('uuid');
 
+// Import the event types
+const gameEvents = require('../assets/json/game-events.json');
+const blackjackEvents = require('../assets/json/blackjack-events.json');
+const pokerEvents = require('../assets/json/poker-events.json');
+const rouletteEvents = require('../assets/json/roulette-events.json');
+
+const gameTypes = require('../assets/json/game-types.json');
 module.exports = class DBHelper {
     constructor(casinoUsers, casinoGames, gameLog, logger) {
         this.casinoUsers = casinoUsers; // collection of casino players
@@ -49,14 +57,15 @@ module.exports = class DBHelper {
         }
       }
     async createGame(gameData) {
+      gameData.gameId = uuidv4();
       const newGame = await CasinoGame.create(gameData);
-      this.casinoGames.set(newGame.gameId, newGame); // Add the new game to the collection
-      return newGame; // Return the newly created game object
+      this.casinoGames.set(newGame.gameId, gameData); // Add the new game to the collection
+      return newGame.gameId; // Return the newly created game object
     }
-    async deleteGame(id) {
+    async deleteGame(id, force) {
       const game = this.casinoGames.get(id);
 
-      if (game) {
+      if (game && force) {
         const affectedRows = await CasinoGame.destroy({ where: { gameId: id } });
 
         if (affectedRows > 0) {
@@ -65,11 +74,27 @@ module.exports = class DBHelper {
         } else {
           throw new Error('No game rows were deleted from the database.');
         }
+      } else if (!force) {
+        this.casinoGames.delete(id);
+        return game;
       } else {
         return null; // Indicate game not found
       }
     }
     async createGameLog(gameLogData) {
+      switch (gameLogData.gameType) {
+        case gameTypes.BLACKJACK:
+          gameLogData.gameEvent = blackjackEvents;
+          break;
+        case gameTypes.POKER:
+          gameLogData.gameEvent = pokerEvents;
+          break;
+        case gameTypes.ROULETTE:
+          gameLogData.gameEvent = rouletteEvents;
+          break;
+        default:
+          gameLogData.gameEvent = 'INVALID_GAME_TYPE';
+      }
       const newGameLog = await CasinoGameLog.create(gameLogData);
       return newGameLog; // Return the newly created game log object
     }
